pack event components in 1 file and relocate
