--- client/src/pages/events/components/EventPagination.js
@@ -1,60 +1,0 @@
-import { useState, useEffect } from "react";
-import PropTypes from "prop-types";
-import { Box, Pagination } from "@mui/material";
-
-function EventPagination({ onPageChanged, page, pageSize, total }) {
-	const [paginationSize, setPaginationSize] = useState("large");
-	const [screenSize, getDimension] = useState({
-		dynamicWidth: window.innerWidth,
-	});
-
-	const setDimension = () => {
-		getDimension({
-			dynamicWidth: window.innerWidth,
-		});
-	};
-
-	useEffect(() => {
-		window.addEventListener("resize", setDimension);
-		if (screenSize.dynamicWidth > 960) {
-			setPaginationSize("large");
-		} else if (screenSize.dynamicWidth > 640) {
-			setPaginationSize("medium");
-		} else {
-			setPaginationSize("small");
-		}
-		return () => {
-			window.removeEventListener("resize", setDimension);
-		};
-	}, [screenSize]);
-
-	return (
-		<Box className="pagination-box">
-			<Pagination
-				size={paginationSize}
-				count={Math.ceil(total / pageSize)}
-				page={page}
-				onChange={onPageChanged}
-				siblingCount={2}
-				showFirstButton
-				showLastButton
-				color="primary"
-				shape="rounded"
-				sx={{
-					".MuiPaginationItem-root": {
-						color: "#fff",
-					},
-				}}
-			/>
-		</Box>
-	);
-}
-
-export default EventPagination;
-
-EventPagination.propTypes = {
-	onPageChanged: PropTypes.func.isRequired,
-	page: PropTypes.number.isRequired,
-	pageSize: PropTypes.number.isRequired,
-	total: PropTypes.number.isRequired,
-};
