--- client/src/pages/events/components/EventsHeader.js
@@ -1,11 +1,0 @@
-import "../events.css";
-
-function EventsHeader() {
-	return (
-		<div className="events-header">
-			<h1>Events</h1>
-		</div>
-	);
-}
-
-export default EventsHeader;
